Suggestion for admin table
The current columns:
When | 404 Path | Came From | IP Address | User Agent | Custom Redirect

The proposed changes:
**Date** | 404 Path | **From** | IP Address | User Agent | **Redirect**

Reasons:
**Date** is a wider used term, direct.
**From** is simple, shorter and direct.
**Redirect** is shorter, direct, and comply with the below changes I've also made to the content of the column.
For the **Redirect** column data, the data isn't in the hover title but actually listed, wich is better for sorting purposes and can be read right away, without hovering. The current **Update** term isn't also correct, might be better something like "Customize" as the options are **Default** or the _customized URL_. So I've kept the hover text with "Customize".